Added in a bunch of data

+++ ThemeParkDatabase/Models/MaintenanceRequest.cs
@@ -10,7 +10,7 @@ namespace ThemeParkDatabase.Models
         public DateTime DateRequested { get; set; }
         public DateTime? DateResolved { get; set; }
         public string CurrentStatus { get; set; }
-        public decimal EstimatedCost { get; set; }
+        public int EstimatedCost { get; set; }
         public int AttractionId { get; set; }
 
         public  Attraction Attraction { get; set; }
